BUG: fix bug in simulate(::DiscreteMarkovProcess, ...)
The old code used the markov state indices as a _linear index_
to extract values. This was incorrect as it would only ever access the
first column of the processes indices.

In this routine I separated the indices and values functionality into
two functions: simulate and simulate_values. I don't think we were ever
actually asking for the values to be returned instead of the indices,
so I don't think this is much of a problem.

Also, I switched to using QuantEcon's routines for simulating the
indices for two reasons:

1. It is about 10x faster
2. We (I) don't have to maintain/test two sets of routines that do that
operation